config/blog/requirements.txt
Django-4.0.4